+++ CHANGELOG.md
@@ -724,3 +724,26 @@
   - core/results/event_data_14222852410.json
   - version.txt
 ---------------------------------
+## Version 0.0.46
+- **Branch**: refs/pull/267/merge
+- **Commit**: [Ka0S] Uploading resume execution files to the repository...
+- **Author**: Jhony Chaves
+- **Date**: Thu Apr  3 08:02:27 UTC 2025
+- **Message**: H220 web de resultados
+- **Files Changed**:
+  - core/dashboard/dashboard/sections/backLogs.json
+  - core/dashboard/js/charts.js
+  - core/dashboard/js/dashboard.js
+  - core/dashboard/js/loaders.js
+  - core/dashboard/js/navigation.js
+  - core/dashboard/js/sections.js
+  - core/dashboard/js/workflows.js
+  - core/dashboard/templates/backLogs.html
+  - core/dashboard/templates/index.html
+  - core/results/event_data_14220989170.json
+  - core/results/event_data_14221470395.json
+  - core/results/event_data_14236187098.json
+  - core/results/event_data_14236638237.json
+  - core/results/event_data_14237927203.json
+  - version.txt
+---------------------------------